requirements.txt
torch==1.7.1
torchvision==0.8.2
PyYAML==5.3.1
Pillow==8.1.2
tensorboard==2.4.1
tensorboardX==2.1
tqdm==4.47.0
Flask==1.1.2
beautifulsoup4==4.9.3
werkzeug==1.0.1